+++ src/OpenTelemetry.AutoInstrumentation/Configurations/LogSettings.cs
@@ -1,8 +1,10 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using OpenTelemetry.AutoInstrumentation.Configurations.FileBasedConfiguration;
 using OpenTelemetry.AutoInstrumentation.Configurations.Otlp;
 using OpenTelemetry.AutoInstrumentation.Logging;
+using Vendors.YamlDotNet.Core.Tokens;
 
 namespace OpenTelemetry.AutoInstrumentation.Configurations;
 
@@ -43,7 +45,12 @@ internal class LogSettings : Settings
     /// </summary>
     public OtlpSettings? OtlpSettings { get; private set; }
 
-    protected override void OnLoad(Configuration configuration)
+    /// <summary>
+    /// Gets tracing Batch Processor Configuration.
+    /// </summary>
+    public BatchProcessorConfig BatchProcessorConfig { get; private set; } = new();
+
+    protected override void OnLoadEnvVar(Configuration configuration)
     {
         LogsEnabled = configuration.GetBool(ConfigurationKeys.Logs.LogsEnabled) ?? true;
         LogExporters = ParseLogExporter(configuration);
@@ -51,6 +58,12 @@ internal class LogSettings : Settings
         {
             OtlpSettings = new OtlpSettings(OtlpSignalType.Logs, configuration);
         }
+
+        BatchProcessorConfig = new BatchProcessorConfig(
+            scheduleDelay: configuration.GetInt32(ConfigurationKeys.Traces.BatchSpanProcessorConfig.ScheduleDelay),
+            exportTimeout: configuration.GetInt32(ConfigurationKeys.Traces.BatchSpanProcessorConfig.ExportTimeout),
+            maxQueueSize: configuration.GetInt32(ConfigurationKeys.Traces.BatchSpanProcessorConfig.MaxQueueSize),
+            maxExportBatchSize: configuration.GetInt32(ConfigurationKeys.Traces.BatchSpanProcessorConfig.MaxExportBatchSize));
 
         IncludeFormattedMessage = configuration.GetBool(ConfigurationKeys.Logs.IncludeFormattedMessage) ?? false;
         EnableLog4NetBridge = configuration.GetBool(ConfigurationKeys.Logs.EnableLog4NetBridge) ?? false;
@@ -62,6 +75,51 @@ internal class LogSettings : Settings
         EnabledInstrumentations = configuration.ParseEnabledEnumList<LogInstrumentation>(
             enabledByDefault: instrumentationEnabledByDefault,
             enabledConfigurationTemplate: ConfigurationKeys.Logs.EnabledLogsInstrumentationTemplate);
+    }
+
+    protected override void OnLoadFile(Conf configuration)
+    {
+        if (configuration.LoggerProvider != null &&
+            configuration.LoggerProvider.Processors != null &&
+            configuration.LoggerProvider.Processors.TryGetValue("batch", out var batchProcessorConfig))
+        {
+            BatchProcessorConfig = batchProcessorConfig;
+            LogsEnabled = true;
+            var logExporters = new List<LogExporter>();
+            var exporters = batchProcessorConfig.Exporter;
+            if (exporters != null)
+            {
+                if (exporters.OtlpGrpc != null)
+                {
+                    logExporters.Add(LogExporter.Otlp);
+                    OtlpSettings = new OtlpSettings(OtlpSignalType.Logs, exporters.OtlpGrpc);
+                }
+
+                if (exporters.OtlpHttp != null)
+                {
+                    logExporters.Add(LogExporter.Otlp);
+                    OtlpSettings = new OtlpSettings(OtlpSignalType.Logs, exporters.OtlpHttp);
+                }
+
+                if (exporters.Console != null)
+                {
+                    logExporters.Add(LogExporter.Console);
+                }
+
+                LogExporters = logExporters;
+            }
+        }
+        else
+        {
+            LogsEnabled = false;
+        }
+
+        var logsInstrumentations = configuration.InstrumentationDevelopment?.DotNet?.Logs;
+
+        if (logsInstrumentations != null)
+        {
+            EnabledInstrumentations = logsInstrumentations.GetEnabledInstrumentations();
+        }
     }
 
     private static IReadOnlyList<LogExporter> ParseLogExporter(Configuration configuration)